Add: parse-color, make put-string use it
Adds terminal/parse-color, which allows for colors
outside of the constants previously used.

It allows for indexed and RGB modes, which were
previously inacessible.

It doesn't change the behavior of existing code,
except in one edge case of someone using put-string
:fg or :bg options with a string, in which case it
now throws an error rather than a default value.

Finally, it includes an escape hatch that allows
for user-provided TextColors to flow through it.